Use SHRT_MAX instead INT_MAX as maximum tooltip width
This fixes problem with empty tooltips on High-DPI displays.

Based on TortoiseSVN rev. 27836.

+++ src/TortoiseMerge/BaseView.cpp
@@ -2960,7 +2960,7 @@ INT_PTR CBaseView::OnToolHitTest(CPoint point, TOOLINFO* pTI) const
 		// we want multi line tooltips
 		CToolTipCtrl* pToolTip = AfxGetModuleThreadState()->m_pToolTip;
 		if (pToolTip->GetSafeHwnd())
-			pToolTip->SetMaxTipWidth(INT_MAX);
+			pToolTip->SetMaxTipWidth(SHRT_MAX);
 
 		return (textrect.PtInRect(point) ? 1 : 2);
 	}
